packet: fix the packet number type to s64
As it says in rfc9000#name-packet-numbers, the packet number is an integer
in the range 0 to 2^62-1. If the packet number for sending reaches 2^62-1,
the sender MUST close the connection without sending a CONNECTION_CLOSE
frame or any further packets.

So change the packet number type from u32 to s64.

Currently, it keeps the number length as 4 in the packet header, when the
number > U32_MAX, as it says in rfc9000#name-packet-number-encoding-and-,
the number of bits required to represent the packet number is reduced by
including only the least significant bits of the packet number.

Note this patch was tested by setting QUIC_PN_MAP_BASE_PN to different
values on both server and client sides.

+++ net/quic/cong.c
@@ -12,7 +12,7 @@
 
 #include "socket.h"
 
-static void quic_reno_cwnd_update_after_timeout(struct sock *sk, u32 packet_number, u32 transmit_ts)
+static void quic_reno_cwnd_update_after_timeout(struct sock *sk, s64 packet_number, u32 transmit_ts)
 {
 	struct quic_packet *packet = quic_packet(sk);
 	struct quic_cong *cong = quic_cong(sk);
@@ -61,7 +61,7 @@ static void quic_reno_cwnd_update_after_timeout(struct sock *sk, u32 packet_numb
 	quic_outq_set_window(quic_outq(sk), cong->window);
 }
 
-static void quic_reno_cwnd_update_after_sack(struct sock *sk, u32 acked_number, u32 transmit_ts,
+static void quic_reno_cwnd_update_after_sack(struct sock *sk, s64 acked_number, u32 transmit_ts,
 					     u32 acked_bytes)
 {
 	u32 inflight = quic_outq_inflight(quic_outq(sk));
@@ -125,12 +125,12 @@ static struct quic_cong_ops quic_congs[] = {
 	},
 };
 
-void quic_cong_cwnd_update_after_timeout(struct sock *sk, u32 packet_number, u32 transmit_ts)
+void quic_cong_cwnd_update_after_timeout(struct sock *sk, s64 packet_number, u32 transmit_ts)
 {
 	quic_cong(sk)->ops->quic_cwnd_update_after_timeout(sk, packet_number, transmit_ts);
 }
 
-void quic_cong_cwnd_update_after_sack(struct sock *sk, u32 acked_number, u32 transmit_ts,
+void quic_cong_cwnd_update_after_sack(struct sock *sk, s64 acked_number, u32 transmit_ts,
 				      u32 acked_bytes)
 {
 	quic_cong(sk)->ops->quic_cwnd_update_after_sack(sk, acked_number, transmit_ts,
